scripts/requirements.txt
nba_api==1.11.4
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.11
python-dotenv==1.1.1
requests==2.33.1